agents/tools_agent/requirements.txt
google-adk >= 1.2.1
google-cloud-aiplatform[adk, agent_engines] >= 1.96.0
silly